+++ CHANGELOG.md
@@ -46,6 +46,8 @@
 - Added line to testlog to tell the user to expect warnings as part of normal testing.
 - Fixed GPU refactorization allocation and synchronization errors, uninitialized GMRES example initial guesses, and CUDA ILU0 failures on stored numerical zero pivots.
 
+- Added cuDSS implementation of refactorization
+
 ## Changes to Re::Solve in release 0.99.2
 
 ### Major Features